feat: add unit test
