+++ package.json
@@ -56,7 +56,8 @@
   },
   "jest": {
     "transform": {
-      "^.+\\.js$": "babel-jest"
+      "^.+\\.js$": "babel-jest",
+      ".+\\.(css|styl|less|sass|scss|png|jpg|svg|ttf|woff|woff2|wav)$": "jest-transform-stub"
     },
     "moduleNameMapper": {
       "\\.(css|less)$": "jest-transform-stub"
