+++ src/js/components/utils.js
@@ -1,0 +1,13 @@
+const shuffleArray = (array) => {
+  for (let i = array.length - 1; i > 0; i--) {
+    let j = Math.floor(Math.random() * (i + 1));
+    [array[i], array[j]] = [array[j], array[i]];
+  }
+  return array;
+}
+
+const doubleArray = (array) => {
+  return Array(2).fill(array).flat();
+}
+
+export { shuffleArray, doubleArray };
