+++ src/test/memory.test.js
@@ -1,0 +1,5 @@
+import { doubleArray } from '../js/components/utils';
+
+test('doubleArray', () => {
+  expect(doubleArray([1, 2])).toEqual([1, 2, 1, 2]);
+});
